# api/openapi.yaml
openapi: 3.1.0
info:
  title: Vendor Status Monitoring Control Plane
  version: 1.0.0-m5
  description: Tenant-scoped API for vendor status, subscriptions, endpoints, delivery evidence, audit export, and live canonical events.
servers:
  - url: /
security:
  - bearerAuth: []
  - browserSession: []
paths:
...
  /v1/tenants/{tenant}/session:
    get:
      operationId: getTenantSession
      parameters: [{$ref: '#/components/parameters/Tenant'}]
      responses:
        '200': {description: Tenant-scoped identity and browser CSRF token; supports service accounts, content: {application/json: {schema: {type: object}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
  /v1/tenants/{tenant}/logout:
    post:
      operationId: logoutTenantSession
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/CSRF'}]
      responses:
        '200': {description: "Clears browser session; service-account callers discard their token locally"}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
  /auth/{tenant}/login:
    get:
      security: []
      operationId: startOIDCLogin
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - in: query
          name: issuer
          schema: {type: string, format: uri}
      responses:
        '302': {description: "Redirect to the configured OIDC authorization endpoint"}
...
  /auth/logout:
    post:
      operationId: logoutBrowserSession
      parameters: [{$ref: '#/components/parameters/CSRF'}]
      responses:
        '200': {description: "Session cleared"}
  /v1/tenants/{tenant}/vendors:
    get:
      operationId: listVendors
      parameters: [{$ref: '#/components/parameters/Tenant'}]
      responses:
        '200': {$ref: '#/components/responses/List'}
  /v1/tenants/{tenant}/vendors/{vendor}/status:
    get:
      operationId: getVendorStatus
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - {in: path, name: vendor, required: true, schema: {type: string}}
      responses:
        '200': {description: Aggregated vendor status, content: {application/json: {schema: {$ref: '#/components/schemas/VendorStatus'}}}}
        '404': {$ref: '#/components/responses/Problem'}
  /v1/tenants/{tenant}/sources:probe:
    post:
      operationId: probeSource
      parameters: [{$ref: '#/components/parameters/Tenant'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SourceInput'}}}}
      responses:
        '200': {description: Adapter capabilities without persistence, content: {application/json: {schema: {type: object}}}}
  /v1/tenants/{tenant}/sources:
    get:
      operationId: listSources
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/List'}}
    post:
      operationId: createSource
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SourceInput'}}}}
      responses:
        '201': {description: Source created, content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}}
        '409': {$ref: '#/components/responses/Problem'}
  /v1/tenants/{tenant}/sources/{source}:
    get:
      operationId: getSource
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Source'}]
      responses: {'200': {description: Source, content: {application/json: {schema: {$ref: '#/components/schemas/Source'}}}}}
    patch:
      operationId: updateSource
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Source'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [enabled], properties: {enabled: {type: boolean}}}}}}
      responses: {'200': {description: "Updated source"}}
    delete:
      operationId: deleteSource
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Source'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {description: "Workspace source soft-deleted; history retained. Returns id and deleted=true"}, '404': {description: "Source not owned by workspace"}}
  /v1/tenants/{tenant}/sources/{source}/adapter-rollouts:
    post:
      operationId: createAdapterRollout
      description: Start a bounded shadow rollout for a tenant-private source. Shared public sources remain platform-managed.
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Source'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdapterRolloutInput'}}}}
      responses:
        '201': {description: Shadow rollout started, content: {application/json: {schema: {$ref: '#/components/schemas/AdapterRollout'}}}}
        '409': {$ref: '#/components/responses/Problem'}
  /v1/tenants/{tenant}/sources/{source}/adapter-rollout:
    get:
      operationId: getLatestSourceAdapterRollout
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Source'}]
      responses:
        '200': {description: Most recently created rollout and current statistics, content: {application/json: {schema: {$ref: '#/components/schemas/AdapterRolloutView'}}}}
        '404': {$ref: '#/components/responses/Problem'}
  /v1/tenants/{tenant}/adapter-rollouts/{rollout}:
    get:
      operationId: getAdapterRollout
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Rollout'}]
      responses:
        '200': {description: Rollout and current quality-gate statistics, content: {application/json: {schema: {$ref: '#/components/schemas/AdapterRolloutView'}}}}
        '404': {$ref: '#/components/responses/Problem'}
  /v1/tenants/{tenant}/adapter-rollouts/{rollout}/promote:
    post:
      operationId: promoteAdapterRollout
      description: Atomically promote the candidate only when minimum samples, mismatch rate, and error rate pass.
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Rollout'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdapterRolloutDecision'}}}}
      responses: {'200': {description: "Candidate promoted"}, '409': {$ref: '#/components/responses/Problem'}}
  /v1/tenants/{tenant}/adapter-rollouts/{rollout}/rollback:
    post:
      operationId: rollbackAdapterRollout
      description: Stop a shadow rollout or restore the baseline adapter after promotion.
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Rollout'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdapterRolloutDecision'}}}}
      responses: {'200': {description: "Baseline retained or restored"}, '409': {$ref: '#/components/responses/Problem'}}
  /v1/tenants/{tenant}/incidents:
    get:
      operationId: listIncidents
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - {in: query, name: vendor, schema: {type: string}}
        - {in: query, name: phase, schema: {type: string}}
        - {in: query, name: since, schema: {type: string, format: date-time}}
      responses: {'200': {$ref: '#/components/responses/List'}}
  /v1/tenants/{tenant}/incidents/{incident}:
    get:
      operationId: getIncidentTimeline
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - {in: path, name: incident, required: true, schema: {type: string, format: uuid}}
      responses: {'200': {description: "Incident and ordered update timeline"}}
  /v1/tenants/{tenant}/events/stream:
    get:
      operationId: streamCanonicalEvents
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - {in: header, name: Last-Event-ID, schema: {type: string}, description: Signed cursor returned in the preceding SSE id field}
      responses:
        '200': {description: SSE stream with canonical_event messages, content: {text/event-stream: {schema: {type: string}}}}
  /v1/tenants/{tenant}/subscriptions:
    get:
      operationId: listSubscriptions
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/List'}}
    post:
      operationId: createSubscription
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SubscriptionInput'}}}}
      responses: {'201': {description: "Subscription created"}}
  /v1/tenants/{tenant}/subscriptions/{subscription}:
    get:
      operationId: getSubscription
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Subscription'}]
      responses: {'200': {description: "Subscription"}}
    put:
      operationId: updateSubscription
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Subscription'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SubscriptionInput'}}}}
      responses: {'200': {description: "Subscription updated"}, '409': {$ref: '#/components/responses/Problem'}}
    delete:
      operationId: deleteSubscription
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Subscription'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {description: "Subscription soft-deleted; history retained. Returns id and deleted=true"}}
  /v1/tenants/{tenant}/endpoints:
    get:
      operationId: listEndpoints
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses: {'200': {$ref: '#/components/responses/List'}}
    post:
      operationId: createEndpoint
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointInput'}}}}
      responses: {'201': {description: "Encrypted endpoint created"}}
  /v1/tenants/{tenant}/endpoints/{endpoint}:
    get:
      operationId: getEndpoint
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Endpoint'}]
      responses: {'200': {description: "Endpoint metadata; credentials are never returned"}}
    put:
      operationId: rotateEndpoint
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Endpoint'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EndpointInput'}}}}
      responses: {'200': {description: "Endpoint configuration rotated"}, '409': {$ref: '#/components/responses/Problem'}}
    delete:
      operationId: deleteEndpoint
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Endpoint'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'200': {description: "Endpoint soft-deleted; rule references removed; empty rules disabled; history retained. Returns id and deleted=true"}}
  /v1/tenants/{tenant}/endpoints/{endpoint}/test:
    post:
      operationId: testEndpoint
      parameters: [{$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/Endpoint'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CSRF'}]
      responses: {'202': {description: "Durable asynchronous endpoint test queued"}}
  /v1/tenants/{tenant}/endpoint-tests/{test}:
    get:
      operationId: getEndpointTest
      parameters:
        - $ref: '#/components/parameters/Tenant'
        - {in: path, name: test, required: true, schema: {type: string, format: uuid}}
      responses: {'200': {description: "Endpoint test result"}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT or service-account token}
    browserSession: {type: apiKey, in: cookie, name: statushub_session}
  parameters:
    Tenant: {in: path, name: tenant, required: true, schema: {type: string}, description: Tenant UUID or slug; authentication must resolve to the same tenant}
...
    CSRF: {in: header, name: X-CSRF-Token, schema: {type: string}, description: Required for cookie-authenticated unsafe requests; not required with Authorization bearer tokens}
  responses:
    List:
      description: Cursor-paginated collection
      content: {application/json: {schema: {$ref: '#/components/schemas/Page'}}}
    Problem:
      description: RFC 9457-style problem response
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  schemas:
...
    VendorStatus:
      type: object
      required: [id, slug, name, status, active_incidents, visible_sources]
      properties:
        collection: {$ref: '#/components/schemas/CollectionStatus'}
        id: {type: string, format: uuid}
        slug: {type: string}
        name: {type: string}
        status: {type: string, enum: [operational, degraded, partial_outage, major_outage, under_maintenance, unknown]}
        active_incidents: {type: integer}
        visible_sources: {type: integer}
        last_observed_at: {type: string, format: date-time}
...